refactor: common/* includes working
